xz: Move resetting of string_to_filter_used to forget_filter_chain().
Before this, coder_add_filter() and coder_set_preset() set
string_to_filter_used = false but coder_set_extreme() didn't.
In practice only coder_add_filter() needs to do it as its the
only place that reads the variable. But it's clearer if the
resetting is done in forget_filter_chain().