xz: Refactor duplicated check for custom suffix when using --format=raw
The raw format with xz must have a custom suffix set since
there is no default suffix. This check was previously duplicated
when creating the compressed and uncompressed filename. Now,
the check is done once in the code and will cause a fatal error
since compression/decompression cannot be done so continuing
after this is pointless.