liblzma: Change use of LZMA_STR_NO_VALIDATION in lzma_str_to_filters().
The LZMA_STR_NO_VALIDATION flag name and description is ambiguous to how
it handles individual filters that cannot be used in .xz format
(only lzma1 for now). Now LZMA_STR_NO_VALIDATION is a super set of
the LZMA_STR_ALL_FILTERS flag, so they don't need to be used together.